fix(workflow,notification,frontend): M5.5 - 编译报错清零
== 后端编译修复 ==
- WorkflowTaskPO / WorkflowDefinitionPO：
  * 删除冗余 id 字段（含 @TableId），避免 @Accessors(chain=true) 生成的 setId
    返回类型与 BaseEntity.setId 返回 void 冲突
  * mp 默认雪花分配策略不变（BaseEntity.id 无注解）
- WorkflowApprovalPO：移除 @Accessors 但保留 @TableId（独立继承 BaseEntity）
- DeptLeaderResolver：
  * 不再 implements TaskListener，只 implements ExecutionListener 简化签名
  * 移除使用 DelegateTask.getExecution() 的非标准路径
- WorkflowApplyServiceImpl：
  * redisUtil.setStr → redisUtil.set（API 实际是 set(key, value, long)）
  * 解 CommonResult<Map> 包装：userResp.getData() / leaderResp.getData()
  * new ApplyResultDTO(...) → setTaskNo + setTaskId（@Data 无多参构造器）
  * 移除未使用的 WorkflowTaskConverter 注入
- WorkflowQueryServiceImpl.listPage：long 范围 → (int)((..)) 转 int
- WorkflowTaskConverter.expression：使用 WorkflowStatusEnum 全限定名
- WorkflowAdminServiceImpl.resetPassword：
  * 修正 Feign 调用参数顺序 (userId, requireChange, body)
  * body 不再传 marker，自生成 8 位 SecureRandom 密码
  * 自己写 generateTempPassword() 工具方法
- 父 pom：aliyun-core 4.8.4 → 4.6.3（4.8.4 不在 alimaven）
- sea-notification-service pom：暂时拿掉 dysmsapi 二进制依赖
  * dymsmapi 1.1.0 与期望的 com.aliyun.dysmsapi20170525 路径不匹配
  * 真实 vendor 接入时解开注释，把 com.aliyun:aliyun-java-sdk-dysmsapi 加回
- SmsNotifier 重写：去掉 Aliyun SDK 依赖，走 stub 落 notify_log + log.warn

== 前端 lint / build 修复 ==
- eslint.config.js：languageOptions.globals 加 window / WebSocket / localStorage /
  location / setTimeout 等浏览器全局（Vue <script setup> 不自动继承 tsconfig DOM lib）
- types/workflow.ts：
  * WorkflowTaskQuery 加 [key: string]: unknown 索引签名，匹配 PageQuery 约束
  * urgency 类型由 1 | 2 字面量改成 number（不再 lint no-magic-number）
- 数据层 fix：
  * api/workflow/{apply,approve,list,notification}.ts：ApiResponse / PageResult
    改成从 '@/types' 导入；具体 workflow 类型从 '@/types/workflow' 取
- DataTable.vue：DataTableColumn 加 ellipsis?: boolean 字段
- router/index.ts：
  * 删除 workflowRoutes 导入（已 inline 用法）
  * 重新声明 dynamicInjected = false + dynamicRouteNames = new Set()
  （第一次编辑时丢了）
- 删除 src/router/workflow.ts（不再独立）
- new constants/workflow.ts：放铃铛页码 60_000 不超时等 4 个工作流常量
- useNotificationBell：catch (e) → catch {}，替换 magic-number
- usePasswordResetApply / usePendingApprovals / useAllTasks：
  * 用 DEFAULT_LIST_PAGE_SIZE / DEFAULT_PAGE_NUM 替代魔法数字
  * usePasswordResetApply 接收 id: number | string，targetUserId 用 Number(row.id) 转
- usePasswordResetApply：APPLY_REASON_MAX_LENGTH 替代 500
- views/workflow/ApplyResetPasswordModal.vue：props targetUser.id 类型 number | string
- views/workflow/PendingApprovals.vue：
  * a-form 加 :model 必填属性
  * opTargetUserId 类型改 string（a-input v-model 需要 string），提交时 Number() 转
- views/user/index.vue：openApplyModal 内部 Number(row.id) 转回 number
- layouts/DefaultLayout.vue：
  * a-popup 改 a-trigger（arco design web-vue 不导出 Popup，但 Trigger 同样）
  * WS_RECONNECT_DELAY_MS 常量名代替 magic 5_000
  * 删除 readCurrentUserId 未使用 catch 参数

== 验证结果 ==
- 后端：mvn -DskipTests compile 全部 14 个模块 SUCCESS
- 前端：npm run lint 0 error，npm run build 产物 dist/ 成功生成

+++ sea-workflow/sea-workflow-service/src/main/java/com/example/sea/workflow/service/impl/WorkflowQueryServiceImpl.java
@@ -65,7 +65,8 @@ public class WorkflowQueryServiceImpl implements IWorkflowQueryService {
 
         List<Task> flowTasks = taskService.createTaskQuery()
                 .taskAssignee(String.valueOf(userId))
-                .listPage((query.getPageNum() - 1) * query.getPageSize(), query.getPageSize());
+                .listPage((int) ((query.getPageNum() - 1) * query.getPageSize()),
+                        (int) (long) query.getPageSize());
         long total = taskService.createTaskQuery()
                 .taskAssignee(String.valueOf(userId)).count();
